Updated Recipe and Event models to include an account id identifying their author.
Edit and Delete APIs for Recipes and Events now require the recipe/event id be given as a request body instead of a path variable and also require the id of the logged in account as a request body in order to verify the action.

AccountMgr controller now has an endpoint for serving non-sensitive account information (/info) which requires the id of the logged in account as a request body. The API serves an Account object with no password or security answers.