AccountMgrController: /username, /password, /tags all accept a single Account model as a request body which can hold and pass the necessary information for each function.
All functions expect the Account to contain the id (String) of the logged in account which sent the request.

/username expects the Account to contain an updated email/username (String).

/password expects the Account to contain securityAnswers (List<String>) and an updated password (String).

/tags requires the Account contain a list of tags (List<String>).
------------------------------------

EventMgrController: /edit and /delete now accept a single Event model as a request body which can hold and pass the necessary information for each function.

All functions expect the Event to contain an authorID (String) and eventID (String).

/edit expects updated values for the remainder of the Event's attributes.
------------------------------------

RecipeMgrController: /edit and /delete now accept a single Recipe model as a request body which can hold and pass the necessary information for each function.

All functions expect the Recipe to contain an authorID (String) and recipeID (String).

/edit expects updated values for the remainder of the Recipe's attributes.