Added new attribute to Account model: List<String> recipes This list containes the recipeIDs (Strings) of an Account's favorite Recipes.
/account-mgr/favorites added:
accepts an Account as a request body and uses the accountID (String) and recipes (List<String>) attributes of that Account.
Updates the recipes (List<String>) of the Account associated with the provided accountID (String) to the provided list of recipeIDs (Strings).

/recipe-book/favorites added:
accepts an accountID (String) as a request parameter and serves a list of the Account's favorite recipes (List<Recipe>).