New API endpoint /event-list/join accepts a List<String> as a request body.
It expects the first element of this list to be the event ID, the second element to be the account ID, and the third to be the account username.

The remaining elements are expected to be less than 2048 characters and non-null values.

The order of these elements is up for discussion, but we'd discussed a pattern of recipe title/description, recipe ID, etc.